Add support for a MySQL target.
In this commit there's just the Target time. There's no CopyTask similar to the postgres implementation b/c it's still not needed. Not a big deal adding it later though, they are conceptually the same.
With the current code you could define at output() method which returns a MySqlTarget and on this output you can touch() to mark the task as completed. Luigi will play along and invoke the exists() method to check whether the task is completed or not - like any other target type