fixed two bugs in hive.py
The calls to ctors for HivePartitionTarget and HiveTableTarget put the arguments in the wrong order (database first). The order has been fixed to be table first.

If parent_dir is empty then the test failed. The new code checks to see if there is even a parent_dir to test the existence of.