hdfs: Add unix username to /tmp paths
The idea is that we would rather see paths like

        /tmp/<username>/blah/luigi-tempfile-1234

over just

        /tmp/blah/luigi-tempfile-1234

The (very) common issue that Spotify is experiencing is that users first
run their job as their own user to test their luigi job, then they use
the production user. The production user will then fail to run, because
the owner of the file will be the interactive user and not the
production user. It's a serious waste of time to chowning folders all
the time and encountering quite unnecessary permission errors.