Lessen verbosity of tests
The problem at hand is that the travis logs get truncated to 10'000
lines.  Even though you can download the whole log file, you still get
completely log blind due to the super-verbose output.

The reason we get so much output is because the python logging module
defaults to outputting to stderr which nosetests does not capture. We
want it to capture the output from the loggers, since you only care
about the output if the test fails.

My solution is nice in the sense that it doesn't touch the code at all.
Instead, we configure the logger through the built in way of doing
configuration files in luigi.

I ping @jcrobak since he added client.cfg in `.gitignore` in commit
c642a9ec. I found this use case for commiting client.cfg sensible.

The logging config file was inspired by
https://docs.python.org/2/howto/logging.html