Improve most_common.
No need neither to create a list in items() nor to sort the whole
sequence. We need only the max one, right?

If we have two values with the same frequence than we still have
undefined result as before as dict.items() & dict.iteritems()
return values in an undefined order.