make filesystem inference work for RangeDaily
This brings RangeDaily feature set on par with RangeHourly, by
generalizing infer_bulk_complete_from_fs implementation.

Previously it just was not implemented (as the removed FIXME in
RangeDaily docstring says). It was only implemented for the hourly
case. Scheduling a daily job outputting to e.g. HdfsTarget raised
NotImplementedError.