Remove coverage and coveralls for now
This is about fixing the SUPER-SCARY false positives we get from Travis
CI service. See #759. I see this as urgent, so if the tests fail for
this PR, that's good.

While we haven't had any incident yet, we were about to have two on
Spotify's side and apparently @daveFNbuck have had one already (#752).

Just now, I got some 2.6 error which is still green.
https://travis-ci.org/spotify/luigi/jobs/51362963