Cleans up get_work logic for assistants
Some logic was added to get_work to treat assistants as a special case
throughout the function to deal with the fact that assistants aren't necessarily
stakeholders of the tasks they'll be running. This commit cleans up some of that
logic by performing a check that the worker is either in the stakeholders or is
an assistant and using that in place of all the special cases.

This also cleans up the greedy logic to treat assistants more like other workers.