bumped version to 1.1.1
Just released 0abe27e5c6be616e107745b48dff8939815da73c as 1.1.0 to pypi so we should at least make sure going forward we work on 1.1.1. Unfortunately it's a bit weird that the git history says 1.0.24 where it was really 1.1.0 but whatever.

The reason I didn't release HEAD to pypi was the recent issues with Tornado etc. I want to wait for that to stabilize a bit.