Prevents non-runnable tasks from going to assistants
I was seeing a lot of errors from assistants trying to run external tasks and
being unable to even find and load the class. It turns out assistants were
being treated as potential workers even by tasks with no workers. As a partial
fix, assistants can now only work on tasks that some worker is capable of.

This fix is great for workflows where external tasks only exist to check that
non-luigi jobs are complete, but won't work for tasks that are external to the
assistant but not to all workers.