Does not consider upstream disabled jobs as pending for worker keep-alive
If you schedule a worker with a pending task that depends on a disabled task and
have worker-keep-alive set to true, the worker will stay alive forever waiting
for a pending task that can never be scheduled. This updates the get_work
response not consider upstream disabled jobs as pending so that workers can die
once all potentially runnable jobs are complete.