Keeps a minimal set of tasks alive for assistants
With an assistant, we don't want tasks with no workers to die but we also don't
want to continuously accumulate tasks that we don't need in the scheduler
anymore. This prevents pruning of tasks without stakeholders unless they are
were scheduled as DISABLED or DONE and are not needed by any tasks that are or
will be runnable. This prevents accumulation without reducing functionality.