Improve best case complexity for get_work
Yes, we only improve the best case complexity, not the worst case.
In summary, you'll get sublinear in aomunt of tasks if your worker
doesn't do any tasks that use global resources.

get_work has gotten way to complex once we added global resource
constraints, as the resource constraints interaction with priorities is
a hairy beast requiring a global perspective, not allowing for
worker-local optimization. See these related tickets which added the
expressiveness/complicated stuff:

  * spotify/luigi#445
  * spotify/luigi#462
  * spotify/luigi#496
  * spotify/luigi#513

Note that the optimization once done in spotify/luigi#584 which helped
@daveFNbuck's company a lot doesn't help Spotify's setup. Most of
Spotify's tasks lingering in the scheduler are pending tasks and not
done tasks.

The proposed hacky solution here, which I mostly see as a temporary
solution until we do a bigger rewrite and rethink of the central
scheduler is the following:

Categorize the get_work scenario in two different buckets, "trivial" and
"nontrivial". Trivial is when none of the tasks of the worker require
global resources, in that scenario, just loop over the tasks the worker
have registered for instead of all the tasks. For the non-trivial case,
we do what we've always done previously.

So in that sense this commit contains two parts:

 * Add an "index" so you can quickly retrieve the tasks for a worker
 * Do the special-casing of trivial and nontrivial in get_work

At Spotify, we don't use the resources feature that much, so we're
hoping for tremendous speed improvements, no more 30-second get-work
calls! :)