Adds pypy scheduler tests to tox
I just started running my scheduler with PyPy, and it's reducing latency
so I'd like to stick with it. I've added tests to tox.ini to ensure that
this functionality is not lost with future updates. Psycopg2 isn't
compatible with PyPy, so I moved it out of requirements.txt and into
tox.ini for compatible interpreters. If we wanted to make it work,
psycopg2cffi is compatible with all versions but would require some code
changes.