Implement TaskParameter parameter type
Made for MetaTasks like the Range family. See docs included in this
commit as well.

One vision I have with this commit is to add a TaskParameter `task` to
`core()` and to eventually reintroduce the `--task` parameter. Another
vision is to have a better command line parsing and unify the path for
Task instantiation. This commits is a step in that direction.