Allows non-atomic ftp
Part of my pipeline requires uploading files to an external server via ftp. The
server requires specific file names and rejects and deletes the temporary files
uploaded by RemoteTarget. In order to get around this, I've added an optional
atomic parameter to RemoteTarget.put. By running it with atomic=false, the
the target file is written to directly.