Set upper bounds on packages, assuming semver
Every now and then some package somewhere in internet world changes,
causing our travis builds to fail. The only sensible thing to do is to
not pull the latest. We shouldn't bu unneccesarily strict either. So
assuming others use semver, I copied the current version of everything
that we currently use (based on a contemporary build [1]) and just
bumped the major number and set it as a strictly less-than constraint.

I got motivated to do this since #1103, but this is not the first
occurence.

[1]: https://travis-ci.org/spotify/luigi/jobs/73593492