Allows assistant to work on orphaned tasks
Assistants keep PENDING tasks alive in order to work on them, but aren't
actually able to work on them because we check whether tasks have live workers
before we give them to assistants. This is to prevent assistants from working on
external tasks. In order to allow assistants to work on tasks with no live
workers but not work on external tasks, we store whether a task is runnable in
the scheduler.

To keep compatibility with old pickle files, we treat tasks without runnable set
as runnable. This should minimize the problems caused be the changed schema.