Enables task failure on worker death while assistant is running
We avoid pruning PENDING and RUNNING tasks while an assistant is live, as the
assistant is considered a stakeholder for all runnable tasks. Unfortunately,
the prune function is where we fail RUNNING tasks for which the running worker
has died. So when an assistant is alive, it becomes impossible for a RUNNING
task whose worker has died to ever change status.

This fixes the issue by separating the function to mark RUNNING tasks as FAILED
from the pruning function and ensuring that it gets called regardless of the
presence of assistants.