Unexport private parts of interface.py
The purpose is to make the documentation page pleaner[1] and to not make
it easy for users of luigi to misuse luigi and start relying on things
so we don't dare changing it later.

This will of course break for people who've considered those exports to
be sane to use, but I think that's fine since we're making proper
releases of the luigi package.

[1]: http://luigi.readthedocs.org/en/latest/api/luigi.interface.html