worker: Change wait_interval to FloatParameter
In all uses, the wait_interval is a floating point value. It's also more
natural since it's specifies a duration in seconds. It's also useful for
tests to be able to specify low positive values so tests can run faster.

I also added a guard that wait_interval shouldn't be close to zero, as
the multiprocessing.Queue.get() has an undefined behavior when timeout
is set to zero. For me, it worked on 2.7 but not 3.4, meanwhile Travis
failed both of them.

Since all ints are also floats, this change is backward compatible.