cli: Make --help useful and add --help-all
Before this commit, running `luigi --help` is going to print 300 lines
of probably not useful flags, mostly being about all the combinations of
the registered various versions of the Range flags. Instead, I changed
--help to just print.

    $ ./bin/luigi --help
    usage: luigi [--local-scheduler] [--module CORE_MODULE] [--help] [--help-all]

    optional arguments:
      --local-scheduler     Use a in-memory central scheduler. Useful for testing.
      --module CORE_MODULE  Used for dynamic loading of modules
      --help                Show most common flags and all task-specific flags
      --help-all            Show all command line flags
    Exiting due to --help was passed

And when you're interested in an actual tasks parameter, they will be
displayed when specifying that task.

    $ ./bin/luigi --module examples.foo examples.Bar --help
    ...
    --num EXAMPLES.BAR_NUM, --examples.Bar-num EXAMPLES.BAR_NUM
    ...