Makes computeDepth take linear time
computeDepth takes quadratic time because it needs to find the maximum depth of
each node and this can require repeated traversal as longer paths are found. In
order to compute this in linear time, we instead compute the height of each node
and then subtract this from the height of the root node. Heights can be computed
in linear time by a straightforward recursive algorithm with memoization.

Using a test example with a 400-task dependency chain with each task in the
chain depending on a task with 10000 dependencies the time necessary for
computeDepth on my laptop was reduced from about 5 seconds to about 13ms. This
is in line with the expected 400x speedup from removing quadratic behavior.