Reverts #1296 "Makes computeDepth take linear time"
There was a bug in the assumptions of #1296, as it treated depth and height as
having a 1-1 relationship. This is only true with balanced trees. For example,
consider the following tree:

      A
     / \
    B   C
         \
          D

In this tree, B has height 0 and depth 1 while D has height 0 and depth 2. So
the graph will be laid out weirdly if we compute the depth of B and D by
subtracting their heights from the height of A.

There seems to be another bug in the implementation that can actually throw
B off of the graph, but I'm not sure exactly why that was happening.