Displays truncated graph nodes in visualiser
When the scheduler sends a truncated graph, some nodes may have dependencies
that are not displayed. In order to not give a false impression to the end
user, these nodes are added back and marked as TRUNCATED status. These nodes
show that more dependencies exist and can be clicked on for more information
but mostly preserve a smaller graph.

In order to distinguish these nodes from UNKNOWN status nodes, I added UNKNOWN
and TRUNCATED to the legend. UNKNOWN remains black while TRUNCATED is given a
light purple in the svg graph. In the D3 graph TRUNCATED is not given a color.