Distingiush local and global cli params [#1304]
Since the command line overhaul in #1182. One unexpected behavior
was introduced, namely that --x and --MyTask-x would mean the same
thing. As a consequence, `--x 123` would set parameter x to 123 not only
for the parsed root Task specified on the command line. But for all
parmeters x of that task family. Since even Dave from Houzz was confused
by this, we decided to revert that behavior in issue #1304.

While at it. I also made these changes:

 * I decided to do a quite big refactoring. Including finally
   removing the `is_bool` kwarg from the Parameter constructor.
 * I added a test for deps.py script, as I've already been close to
   breaking it twice.