Allows disabling of worker e-mails via luigi.cfg
After a night of excessive failures, I found that my workers were dying
whenever a job failed and they weren't able to access the overloaded
smtp server. When I tried to find the configuration option to disable
emails and get things running again, I found that there wasn't one,
although error-email falsely claims to be ablo to do that in the
documentation.

I've updated the documentation for error-email to clarify that e-mails
will still be sent for jobs that have owners and added a "none" option
to email-type that disables email sends. I also cleaned up the logging
for why emails are disabled.