Fixes broken redshift test
The redshift test asserts that the output is only called once in run,
when run clearly called it twice. I fixed that by storing the output in
a local variable rather than calling it again.

The redshift test asserted that RedshiftTarget is called with arguments
table_type and sql that it does not accept, so I removed those. There
was also a typo in the update_id argument.

The redshift test verified the existence check by asserting that it was
the only redshift call. I fixed that by changing the assertion to verify
that the existence check was made, as the run makes two other calls
afterward.

It seems that this test is incomplete, but it at least no longer fails.