cli: Mention incorrect taskname over missing args
Currently, if you misspell your task in a invocation like this:

  luigi MyTaskk --my-param 123

Luigi will complain that `--my-param 123` isn't recognized. While that
is technically correct as well, it's more user friendly to first control
that the user entered a correct task name.

It was too hard implementation wise to get this working without any
change to current behavior.  Instead we say that the return code of
luigi is undefined whenever a task is misspelled.  This will only happen
when using luigi cli interactively, assuming that nobody relied on the
exit code when consciously misspelling a task name.