Fix deadlock on Ctrl+C when parallel scheduling
When you do Ctrl+C during the (parallel) scheduling phase, luigi would
previously dead lock and keep pinging the central scheduler and hold the
lock-file. The details are elaborated detail in the code and tests.

This bug can also be verified manually (in addition to the added unit
test). Just run

  rm -r /tmp/bar/  # if needed
  PYTHONPATH=. luigi --module examples.foo examples.Foo --local-scheduler

And do a Ctrl+C on your keyboard while it's checking the completeness
for any of the Bar tasks.

I think this bug was introduced in spotify/luigi#1089.