Make the Worker a context manager
This way, we will shut down the KeepAliveThread even when there is an
error during the scheduling phase.  And context managers is a more
pythonic construct than the invented stop() method. This was also a now
removed TODO in the code.

I also put the side effect of starting the KeepAliveThread into the
__enter__ block. I found it surprising that constructing a Worker()
would start a new thread.

A lot of test cases had to be rewritten. While at it, I cleaned up some
of the tests and also added many forgotten cleanups. Hopefully tests are
more stable from now on.