Integrate webhdfs with rest of hdfs module
With this patch you can use normal HdfsTargets together with webhdfs.
The old WebHdfsTarget still exists but is deprecated.

See the documentation for the [hdfs] section for enabling webhdfs.

This will change the behavior of the existing implementation a bit.
The issue was that it was using [hdfs]namenode_port for the webhdfs
port. But the namenode port that for example snakebite and
hadoopcli uses is different from the webhdfs port.