Allow scheduler to listen on unix socket normally
Previously, the scheduler would only listen on a supplied unix
socket when run in the background.  When run normally, it would
ignore the given unix socket and serve on the default port.  This
is fixed by passing the given unix_socket to server.run when
the scheduler is run normally.