Fix issue with complete external tasks having missing deps
There are certain scenarios where an external task has actual deps,
such as when a task has been made external with externalize() method,
or if someone wants to simply subclass then set run = NotImplemented.

At runtime if the output exists for the external task but not for its
upstream dependencies, this should be OK (task success). However this
case was broken because we failed in the missing-dependencies check
before ever checking task.compete().