Format classes will not read whole files in memory
If an input pipe did not provide a file descriptor via a fileno()
method (such as a S3 key object), the whole input data was read in
memory to be written to a temporary file. To avoid using too much
memory and improve efficiency, the data is copied immediately by
chunks of the size of the file system blocks. This does not affect
most cases of Format classes which pass a file descriptor to the
subprocess.

Fixes issue #1467.