Fixes server-side filtering of tasks in scheduler
Server-side filtering of tasks broke when task ids changed from the string
representation of a Task object to a truncated, cleaned up combination of
parameter values and a partial md5 hash. To fix this, we reconstruct a pretty
id more similar to the old value that was matched against for server-side task
filtering.