Optionally hides DONE nodes in dependency graph
Often when viewing the dependency graph, I'm trying to figure out why a
particular task hasn't run yet. This can be difficult due to the presence of
hundreds or even thousands of DONE tasks cluttering up the view. In order to aid
debugging, this adds a "Hide Done" checkbox to the dependency graph view in the
visualiser. When checked, the scheduler will ignore DONE tasks (other than the
root) when doing graph traversal (including inverse graph traversal).