Speeds up get_work with a number of small tweaks
Some profiling of get_work under different scheduler-state scenarios
showed some obvious areas for improvement without major algo changes.

The biggest contributor was in _upstream_status where the task_stack was
being re-allocated rather than appended-to (_upstream_status is used in
other places as well like task_list).

A remaining area for optimization is the sorting of pending tasks that
happens for each call; this would be a bigger change.