parameters: add --no-FOO syntax for negating a BoolParameter
Add two command-line arguments for BoolParameter's instead of one,
specifically one to set it to True (e.g, --foo) and another (--no-foo)
to set it to False.

For cases where an inverted argument does not make sense, the caller
should set omit_inverted_argument=True on the BoolParameter. For example,
this has been done for --help and --help-all options, as --no-help does
not make any sense.

This allows BoolParameter's to have a default of True and still be settable
to False.

Other notes:
* Removes Parameter._parser_action as moot.
* Merges Parameter._parse_or_no_value into the sole call site
  and changes the logic to account for None in the value check.
* Several tests initially failed with this change because argparse considered
  --n ambiguous with the --no-FOO style options. This is a known issue with
  argparse. These options have been renamed to allow the tests to pass.
  See https://bugs.python.org/issue19814 for details.