assistants bugfix: Also prune UNKNOWN tasks
I'll only motivate this lightly as I assume the previous mistake was a
(serious) oversight.  This bug is a bit hard to realize, because when
you visit the scheduler in the browser the UNKNOWN tasks are not shown
so you don't see them accumulating.  You only notice that the page loads
slower for each day.

In our scheduler instance, there are currently 90% UNKNOWN tasks, most
of which would be pruned after this patch is applied. 90% because its
current uptime is 10 days. The 9 previous days have almost exclusively
UNKNOWN tasks meanwhile today has almost exclusively known tasks.