Add GUI-visible status messages for tasks
This patch adds a primitive for set a "status message" for tasks that re being run(). The message will be visible from the web browser in the central scheduler page.

As was realized during the review for this patch, the multi-argument run() introduced in previous spotify/luigi#1298 doesn't scale, so it got deprecated with this patch.