Adds host resources in addition to global resources
Host resources are resources used on a specific host. Tasks with host_resource
set will be limited by how many can run on the same machine, but there is no
limit to how many can run globally. For example, if you have a process that
requires 2gb of memory, you might set its host_resources to {'memory': 2} and
then set the host resource for memory on each machine according to how much it
has available. If you have a machine with 8gb of memory, you can run 4 such
tasks on it. If you have another machine with 16gb of memory, you can run
another 8 tasks on it.