Pass through parameters on Range tasks to of tasks
Parameters to tasks wrapped by RangeBase and its subclasses can now be
passed through the Range to the 'of' task. This both makes it easier to
specify parameters to tasks running inside of ranges in Python code and
allows for Luigi to differentiate between multiple Range instances
running with the same 'of' parameter.

closes #1627