Modified bulk_complete invocations to pass of_params
In RangeDaily and RangeHourly, the bulk_complete parameter was invoked
with a list of the date/hour parameters to be used for the of task. This
prevents of_params from being provided to the bulk_complete code.
Instead, we change the invocation of the bulk_complete method to use a
partial with the of_params values curried. Thus, any code using the cls
parameter of bulk_complete can behave as normal while still allowing the
of_params to be passed through.

For code that would rather subclass and override missing_datetimes, two
new helper methods are provided: datetime_to_parameters and
parameters_to_datetime. These are parallel to datetime_to_parameter and
parameter_to_datetime and rather than handling only the first argument
to the of class, handle all the required arguments.