raise an error when no columns are found
The method `create_table` uses the length of the first element of the `columns` property to determine how column names are formatted. Unfortunately this is not exhaustive case coverage. For instance, if a `columns` property is formatted improperly, `create_table` can fail silently, matching neither the first nor the second if clause. In such cases we should alert the user that something has gone awry.

To remedy this, here we raise a `ValueError` when neither if clause is met. This both alerts the user to the cause of the problem and halts the application from making erroneous and incorrect progress. If we prefer to make the `CREATE` query optional, we could instead log an error to stderr, however it seems unlikely that we would ever wish to not create our tables.