Caches get_autoconfig_client on a per-thread basis
If too many different autoconfig clients are open at the same time, it can
create too many open files for the operating system to handle. To avoid
this issue, we can simply cache the autoconfig client to ensure that the
same one is being used every time. However, the autoconfig client is not
thread safe so we must have a different one for each thread.

This resolves #1740