Marking as minimum upstream severity instead of max (#1789)
Luigi marks a wrapper task as `UPSTREAM_FAILED` or `UPSTREAM_DISABLED`
when ANY of its upstream task is `FAILED` or `DISABLED`. Moreover, when
the wrapper task is marked as `UPSTREAM_DISABLED`, luigi shut the worker
down as it should do. This causes another PENDING_TASK in the wrapper
task not to run because worker is down. This behaviour should change.
Marking as `UPSTREAM_` should be ALL instead of ANY