Add NumericalParameter (#1799)
Added a new Parameter class, NumericalParameter, that validates that the specified input parameter, an int or float, is in the specified range. The comparison functions, i.e. <= or < may be overridden, and the min and max values are adjustable.

This Parameter type is helpful to reduce repetitive code, such as repeatedly checking that various input numeric parameters are in the permissible ranges. I argue it is also cleaner to specify the constraints on the parameter when it is defined in a Task class.