Allows breaking resource strings on commas in visualizer (#1828)
Using JSON.stringify produces a compact string with no spaces between
separators. When used with a longer resources dict, it can cause the
cell in the task list datatable to become too wide and push the edge of
the table off of the screen. By adding a space after the comma
separators, the resources dict can be broken up and everything fits
comfortably within my laptop's smaller window size again.