Update local locking logic
This patch is made to address some issues with local locking, that is
the locking that deals with the .pid files typically in each machines
/tmp/luigi folder.

There's one bugfix (can massively improve startup performence):

  * Don't keep old entries in the .pid files

There's one change in logic:

  * When running with --take-lock, only allow for one "additional" task
    being run, the previous behavior was that --take-lock tasks always
    could keep running. With this new behavior, it could perhaps be
    possible to always cron your lines with --take-lock.

Lastly, there was much refactoring of the code. Although longer, I find
the readability to have drastcially improved.