Prevents batch tasks from running before their dependencies (#1846)
When adding tasks to an existing batch in get_work, we neglected to check
whether the tasks were schedulable. This meant that in production I found
my pipelines running batches that included jobs with PENDING dependencies.
In order to fix this, we simply add a check that the task is schedulable.

Note that even though self._schedulable(task) gets called twice in the
get_work loop, they fall under different branches of whether best_task is
defined, so this shouldn't slow get_work down much.